+++ CHANGELOG.md
@@ -24,6 +24,8 @@
    `Quantity`.
  * [#11](https://github.com/iliekturtles/uom/issues/11) Add floating point `powi` method for
    `Quantity`.
+ * [#7](https://github.com/iliekturtles/uom/issues/7) `Volume` quantity added. Additional units
+   still need to be added.
 
 ## [v0.14.0] — 2017-05-30